Move formatNumber() to format utils lib
