+++ web-registry/src/components/organisms/ProjectCreditBatchesTable.tsx
@@ -5,8 +5,7 @@ import {
   ActionsTable,
   RenderActionButtonsFunc,
 } from 'web-components/lib/components/table/ActionsTable';
-import { formatNumber } from 'web-components/lib/components/table';
-import { formatDate } from 'web-components/lib/utils/format';
+import { formatDate, formatNumber } from 'web-components/lib/utils/format';
 
 import { truncate } from '../../lib/wallet';
 import { getAccountUrl } from '../../lib/block-explorer';
